error handling

+++ frontend/bugcrud.js
@@ -26,6 +26,13 @@ async function postBugData(){
     const bugDescription = document.getElementById("bug-description").value;
     const prior = prioritySelect.value;
     const url = "/api/bugs";
+    if (!bugTitle.trim() || !bugDescription.trim()){
+        document.getElementById("create-message").innerHTML = "Enter valid values for bug data";
+        setTimeout(() => {
+            document.getElementById("create-message").innerHTML = ""
+        }, 3000);
+        return;
+    }
     const bugobj = {
         title : bugTitle,
         description : bugDescription,
@@ -41,9 +48,16 @@ async function postBugData(){
             },
             body: JSON.stringify(bugobj)
         });
-
         const result = await response.json();
+        if (response.ok) {
+            document.getElementById("create-message"),innerHTML = "Bug successfully created";
+        } else {
+            document.getElementById("create-message").innerHTML = "Bug was not created";
 
+        }
+        setTimeout(() => {
+            document.getElementById("create-message").innerHTML = "";
+        })
     } catch (error) {
         console.log("Error: ", error);
     }
@@ -52,6 +66,13 @@ async function postBugData(){
 async function getBugData(){
     const bugID = document.getElementById("bug-id-get").value;
     const url = `/api/bugs/${bugID}`;
+    if(!bugID || isNaN(bugID)) {
+        document.getElementById("get-message").innerHTML = "Enter a valid bug ID";
+        setTimeout(() => {
+            document.getElementById("get-message").innerHTML = ""
+        }, 3000);
+        return;
+    }
     try {
         const response = await fetch(url, {
             method : "GET",
@@ -60,6 +81,19 @@ async function getBugData(){
             }
         });
         const result = await response.json();
+        if (response.ok) {
+            let returnedobj = "";
+            for (const[key, value] of Object.entries(result)){
+                returnedobj += `${key} : ${value}<br>`;
+            }      
+            document.getElementById("get-message").innerHTML = returnedobj;     
+        } else {
+            document.getElementById("get-message").innerHTML = result.detail;
+
+        }
+        setTimeout(() => {
+            document.getElementById("get-message").innerHTML = "";
+        }, 3000);
     } catch (error) {
         console.log("Error : ", error);
     }
@@ -68,6 +102,13 @@ async function getBugData(){
 async function deleteBugData(){
     const bugID = document.getElementById("bug-id-delete").value;
     const url = `/api/bugs/${bugID}`
+    if(!bugID || isNaN(bugID)){
+        document.getElementById("delete-message").innerHTML = "Enter a valid bug ID";
+        setTimeout(() => {
+            document.getElementById("delete-message").innerHTML = ""
+        }, 3000);
+        return;
+    }
     try {
         const response = await fetch(url, {
             method : "DELETE",
@@ -76,6 +117,14 @@ async function deleteBugData(){
             }
         });
         const result = await response.json();
+        if (response.ok) {
+            document.getElementById("delete-message").innerHTML = "Bug sucessfully deleted";
+        } else {
+            document.getElementById("delete-message").innerHTML = result.detail;
+        }
+        setTimeout(() => {
+            document.getElementById("delete-message").innerHTML = "";
+        }, 3000);
     } catch (error) {
         console.log("Error : ", error)
     }
@@ -88,6 +137,13 @@ async function updateBugData(){
     const bugStatus = document.getElementById("bug-status-update").value;
     const prior = document.getElementById("priority-select-update").value;
     const url = `/api/bugs/${bugID}`;
+    if(!bugID || isNaN(bugID) || !bugTitle.trim() || !bugDescription.trim() || !bugStatus.trim()){
+        document.getElementById("update-message").innerHTML = "Enter a valid values for updating bug";
+        setTimeout(() => {
+            document.getElementById("update-message").innerHTML = ""
+        }, 3000);
+        return;
+    }
     const bugobj = {
         title : bugTitle,
         description : bugDescription,
@@ -104,6 +160,14 @@ async function updateBugData(){
             body: JSON.stringify(bugobj)
         });
         const result = await response.json();
+        if (response.ok) {
+            document.getElementById("update-message").innerHTML = "Bug sucessfully updated";
+        } else {
+            document.getElementById("update-mesage").innerHTML = result.detail;
+        }
+        setTimeout(() => {
+            document.getElementById("update-message").innerHTML = "";
+        }, 3000);
     } catch (error) {
         console.log("Error : ", error)
     }
